Extract showcased game title&author and buttons in components
