+++ newIDE/app/src/GamesShowcase/ShowcasedGameListItem.js
@@ -2,32 +2,17 @@
 import * as React from 'react';
 import {
   type ShowcasedGame,
-  type ShowcasedGameLink,
 } from '../Utils/GDevelopServices/Game';
 import { Card } from '@material-ui/core';
 import Text from '../UI/Text';
-import { Trans } from '@lingui/macro';
 import { Column, Line, Spacer } from '../UI/Grid';
 import { MarkdownText } from '../UI/MarkdownText';
 import { CorsAwareImage } from '../UI/CorsAwareImage';
-import RaisedButton from '../UI/RaisedButton';
 import TagChips from '../UI/TagChips';
-import Apple from '../UI/CustomSvgIcons/Apple';
-import Window from '../Utils/Window';
-import { sendShowcaseGameLinkOpened } from '../Utils/Analytics/EventSender';
-import GooglePlay from '../UI/CustomSvgIcons/GooglePlay';
-import Steam from '../UI/CustomSvgIcons/Steam';
-import Twitter from '../UI/CustomSvgIcons/Twitter';
-import Instagram from '../UI/CustomSvgIcons/Instagram';
-import Twitch from '../UI/CustomSvgIcons/Twitch';
-import ItchIo from '../UI/CustomSvgIcons/ItchIo';
-import Microsoft from '../UI/CustomSvgIcons/Microsoft';
-import GetAppIcon from '@material-ui/icons/GetApp';
-import PlayArrowIcon from '@material-ui/icons/PlayArrow';
-import YouTubeIcon from '@material-ui/icons/YouTube';
-import FlashOnIcon from '@material-ui/icons/FlashOn';
 import { ResponsiveLineStackLayout } from '../UI/Layout';
 import { useResponsiveWindowWidth } from '../UI/Reponsive/ResponsiveWindowMeasurer';
+import ShowcasedGameTitle from './ShowcasedGameTitle';
+import ShowcasedGameButton from './ShowcasedGameButtons';
 
 const styles = {
   thumbnailImageWithDescription: {
@@ -65,67 +50,6 @@ type Props = {|
   onHeightComputed: number => void,
 |};
 
-const getFullUrl = url =>
-  url.startsWith('/') ? 'https://gdevelop-app.com' + url : url;
-
-const getLinkIconAndLabel = (url: string, type: string) => {
-  // Supported links:
-  if (type === '') {
-    return { icon: <PlayArrowIcon />, label: <Trans>Play or download</Trans> };
-  } else if (type === 'learn-more') {
-    return { icon: null, label: <Trans>Learn more</Trans> };
-  } else if (type === 'download' || type === 'download-win-mac-linux') {
-    return { icon: <GetAppIcon />, label: <Trans>Download</Trans> };
-  } else if (type === 'play') {
-    return { icon: <PlayArrowIcon />, label: <Trans>Play</Trans> };
-  }
-  // Officially supported stores/websites/social medias:
-  else if (type === 'App Store') {
-    return { icon: <Apple />, label: 'iOS' };
-  } else if (type === 'Play Store') {
-    return { icon: <GooglePlay />, label: 'Android' };
-  } else if (type === 'Steam') {
-    return { icon: <Steam />, label: 'Steam' };
-  } else if (type === 'Trailer') {
-    return { icon: <YouTubeIcon />, label: <Trans>Trailer</Trans> };
-  } else if (type === 'itch.io') {
-    return { icon: <ItchIo />, label: 'itch.io' };
-  } else if (type === 'Game Jolt') {
-    return { icon: <FlashOnIcon />, label: 'Game Jolt' };
-  } else if (type === 'Twitter') {
-    return { icon: <Twitter />, label: 'Twitter' };
-  } else if (type === 'Microsoft Store') {
-    return { icon: <Microsoft />, label: 'Microsoft Store' };
-  } else if (type === 'Instagram') {
-    return { icon: <Instagram />, label: 'Instagram' };
-  } else if (type === 'Twitch') {
-    return { icon: <Twitch />, label: 'Twitch' };
-  }
-
-  return { icon: null, label: type };
-};
-
-const LinkButton = ({
-  link,
-  showcasedGame,
-}: {|
-  link: ShowcasedGameLink,
-  showcasedGame: ShowcasedGame,
-|}) => {
-  const { url, type } = link;
-  return (
-    <RaisedButton
-      key={type + '-' + url}
-      primary
-      {...getLinkIconAndLabel(url, type)}
-      onClick={() => {
-        sendShowcaseGameLinkOpened(showcasedGame.title, type);
-        Window.openExternalURL(getFullUrl(url));
-      }}
-    />
-  );
-};
-
 export const ShowcasedGameListItem = ({
   showcasedGame,
   onHeightComputed,
@@ -151,9 +75,6 @@ export const ShowcasedGameListItem = ({
 
   const windowWidth = useResponsiveWindowWidth();
   const hasDescription = !!showcasedGame.description;
-
-  const firstLinks = showcasedGame.links.slice(0, 3);
-  const otherLinks = showcasedGame.links.slice(3);
 
   return (
     <div
@@ -187,14 +108,7 @@ export const ShowcasedGameListItem = ({
           />
           <Line expand>
             <Column expand>
-              <ResponsiveLineStackLayout noMargin alignItems="baseline" expand>
-                <Text noMargin displayInlineAsSpan>
-                  {showcasedGame.title}
-                </Text>
-                <Text noMargin size="body2" displayInlineAsSpan>
-                  <Trans>by {showcasedGame.author}</Trans>
-                </Text>
-              </ResponsiveLineStackLayout>
+              <ShowcasedGameTitle showcasedGame={showcasedGame} />
               {showcasedGame.genres.length ? (
                 <TagChips tags={showcasedGame.genres} />
               ) : null}
@@ -205,35 +119,7 @@ export const ShowcasedGameListItem = ({
                 />
               </Text>
               <Spacer />
-              <ResponsiveLineStackLayout
-                justifyContent="flex-end"
-                noColumnMargin
-              >
-                {firstLinks.map((link, index) => (
-                  <LinkButton
-                    key={index}
-                    link={link}
-                    showcasedGame={showcasedGame}
-                  />
-                ))}
-              </ResponsiveLineStackLayout>
-              {otherLinks.length > 0 && (
-                <>
-                  {windowWidth === 'small' && (<Spacer />)}
-                  <ResponsiveLineStackLayout
-                    justifyContent="flex-end"
-                    noColumnMargin
-                  >
-                    {otherLinks.map((link, index) => (
-                      <LinkButton
-                        key={index}
-                        link={link}
-                        showcasedGame={showcasedGame}
-                      />
-                    ))}
-                  </ResponsiveLineStackLayout>
-                </>
-              )}
+              <ShowcasedGameButton showcasedGame={showcasedGame} />
             </Column>
           </Line>
         </ResponsiveLineStackLayout>
